Added network.html
This is a overall network of all existing patterns (after you select
conditions such as support, aiScore etc.) For example, the starting
pattern is [272, 900] in the picture below, you can expand any node to
see more details.

The difference of this graph and the previous one is: 1. This is
generated initially with all data. (see pic 2 for the whole view) 2.
The sequence of nodes are dimensions, like [2d, 3d, 4d, 5d,…..]. This
is not generated based on confidence, it’s purely based on all patterns
in the pattern list you selected (from the first heatmap view).